UI work
mainControlPanelUI reads from defaults and apply settings
HUDUI correctly reference schooltime
Pausing simulation pauses time and classes so they can pause their Agents.

+++ Assets/Scripts/School/FSM/SOffTime.cs
@@ -8,8 +8,8 @@ namespace SES.School
 {
     public class SOffTime : SSchoolBaseState
     {
-        short simLength;
-        short progressionIndex = 0;
+        int simLength;
+        int progressionIndex = 0;
 
         public override void EnterState(SchoolDayProgressionController progressionController)
         {
